pyproject.toml
[project]
name = "opndet"
version = "0.1.0"
description = "Tiny single-class object detector. OpenVINO 2022 opset compatible. No NMS."
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0",
    "numpy>=1.24",
    "onnx>=1.14",
    "onnxruntime>=1.16",
    "onnxscript>=0.1",
    "pillow>=10.0",
    "opencv-python-headless>=4.8",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "ruff>=0.1"]

[project.scripts]
opndet = "opndet.cli:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
opndet = ["configs/*.yaml"]

[tool.ruff]
line-length = 100
target-version = "py310"
